Rewrite chip classification backend to use pure pytorch
