+++ rastervision/backend/torch_utils/chip_classification/model.py
@@ -1,0 +1,8 @@
+from torchvision import models
+from torch import nn
+
+
+def get_model(model_arch, num_labels, pretrained=True):
+    model = getattr(models, model_arch)(pretrained=True, progress=True)
+    model.fc = nn.Linear(model.fc.in_features, num_labels)
+    return model
