+++ rastervision/backend/torch_utils/object_detection/plot.py
